fix: 🐞 client id 숨기기(#143)
