+++ ios/IssueTracker/IssueTracker/Network/ClientID.swift
@@ -1,0 +1,12 @@
+//
+//  ClientID.swift
+//  IssueTracker
+//
+//  Created by Lia on 2021/06/23.
+//
+
+import Foundation
+
+extension LoginURL {
+    static let clientID = "d" /// 민감한 정보 숨기기
+}
